+++ app/build.gradle
@@ -69,11 +69,9 @@ dependencies {
     //Glide
     implementation 'com.github.bumptech.glide:glide:4.13.2'
 
-    // coil
-    implementation "io.coil-kt:coil:2.2.2"
+
 
     // swagger
-    implementation("io.swagger:swagger-annotations:1.5.21")
 
     //hilt
     implementation "com.google.dagger:hilt-android:2.44"
