+++ app/src/main/java/com/example/geeksfit/data/remote/model/PropertiesXXXX.kt
@@ -2,6 +2,6 @@ package com.example.geeksfit.data.remote.model
 
 data class PropertiesXXXX(
     val email: Email,
-    val password: PasswordX,
+    val password: Password,
     val username: UsernameX
 )
