+++ app/src/main/java/com/example/geeksfit/data/repository/RegistRepository.kt
@@ -1,0 +1,11 @@
+package com.example.geeksfit.data.repository
+
+import com.example.geeksfit.data.remote.RegistApi
+import com.example.geeksfit.data.remote.model.Password
+import com.example.geeksfit.data.remote.model.PropertiesXXXX
+import com.example.geeksfit.data.remote.model.Register
+
+class RegistRepository(val registApi: RegistApi) {
+
+    suspend fun register(username: String, email: String,password: String){}
+}
