+++ app/src/main/res/layout/fragment_new_password.xml
@@ -12,7 +12,7 @@
     <ImageView
         android:id="@+id/RgNavigateBack"
         android:layout_width="19dp"
-        android:layout_height="26dp"
+        android:layout_height="30dp"
         android:layout_marginStart="14dp"
         android:layout_marginLeft="14dp"
         android:layout_marginTop="5dp"
