+++ app/src/main/res/layout/fragment_register.xml
@@ -11,10 +11,11 @@
 
     <ImageView
         android:id="@+id/RgNavigateBack"
-        android:layout_width="wrap_content"
+        android:layout_width="19dp"
         android:layout_height="30dp"
         android:layout_marginStart="14dp"
         android:layout_marginLeft="14dp"
+        android:layout_marginTop="5dp"
         android:src="@drawable/navigate" />
 
     <TextView
